Link nfct and helper modules with `-z lazy`
Some distributions, such as Gentoo and Chrome OS, try to link all
programs with `-z now` as a security hardening measure.  This breaks
nfct, because nfct cannot satisfy all of the helper modules' symbols.
Therefore nfct implicitly depends on lazy binding.

Have autoconf probe the linker to see if `-z lazy` works, and if so,
use it to link nfct and the helpers.

conntrackd itself is unaffected, and should still work with `-z now`.

Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>