log: introduce a mechanism to know if log was initialized
This will allow to call dlog() function from all the points in the
execution at runtime.

If the log was not initialized, then we just fprintf and return.

By now, we can't init the log engine earlier because we require config
from the user, so there is a egg-chicken problem.
This means that we can't log parsing messages to logfiles but only to
stderr/stdout.

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>